backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
qdrant-client==1.7.0
openai==1.3.0
pypdf==3.17.1
python-docx==1.1.0
# Use standard PyTorch
torch==2.3.0
sentence-transformers==2.7.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
numpy==1.26.4
email-validator==2.1.0
umap-learn==0.5.5
hdbscan
requests==2.31.0
pdf2image==1.17.0
celery==5.3.6
redis==5.0.1
minio==7.2.0
